requirements.txt
aiohttp~=3.8.1
trafaret=2.1.0
PyYAML=6.0
sqlalchemy=1.4.32
aiopg=1.3.3
pytest=7.1.0
